Time: 29 ms (69.37%), Space: 31.1 MB (53.93%) - LeetHub

+++ 724-find-pivot-index/724-find-pivot-index.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int pivotIndex(vector<int>& nums) {
-        long long int sum=0;
+        int sum=0;
         for(int i=0;i<nums.size();i++)  sum+=nums[i];
         
         int lsum=0;
